move bitwise operators to int methods
